Update README.md
The project “The Radar System” is based on simple hardware as well as software. The components are very simple and handy, which can easily demonstrate the concept of the radar system used widely at military bases, in civil aviation departments, and in other fields.  
The Arduino is coded very simply so that it can be easily understood by the users and can be easily edited accordingly with the corresponding hardware. The case study discusses the use of simpler components to build the abstraction. In addition, this report will also discuss the combination of simpler components with the greatest output.